Adding policy tag test to seperate file
Thinking to add post migration tests seperatly which includes usecases like -
1. post migration cpu/ram/disk check
2. post migration tag check
3. post migration retirement check, etc, etc
