+++ cfme/tests/v2v/test_v2v_migrations.py
@@ -2,7 +2,6 @@
 import fauxfactory
 import pytest
 
-from cfme.base.login import BaseLoggedInPage
 from cfme.fixtures.provider import (dual_network_template, dual_disk_template,
  dportgroup_template, win7_template, win10_template, win2016_template, rhel69_template,
  win2012_template, ubuntu16_template, rhel7_minimal)
@@ -13,7 +12,7 @@ from cfme.markers.env_markers.provider import ONE_PER_VERSION
 from cfme.utils.appliance.implementations.ui import navigator
 from cfme.utils.generators import random_vm_name
 from cfme.utils.log import logger
-from cfme.utils.wait import TimedOutError, wait_for
+from cfme.utils.wait import wait_for
 
 pytestmark = [
     pytest.mark.ignore_stream('5.8'),
@@ -551,7 +550,7 @@ def test_migration_special_char_name(request, appliance, v2v_providers, host_cre
     # wait until plan is in progress
     wait_for(func=view.plan_in_progress, func_args=[migration_plan.name],
         message="migration plan is in progress, be patient please",
-        delay=5, num_sec=3600)
+        delay=5, num_sec=1800)
 
     view.switch_to("Completed Plans")
     view.wait_displayed()
@@ -638,72 +637,3 @@ def test_migration_long_name(request, appliance, v2v_providers, host_creds, conv
     assert view.migration_plans_completed_list.is_plan_succeeded(migration_plan.name)
     migrated_vm = get_migrated_vm_obj(vm_obj, v2v_providers.rhv_provider)
     assert vm_obj.mac_address == migrated_vm.mac_address
-
-
-@pytest.mark.parametrize(
-    "form_data_vm_obj_single_datastore", [["nfs", "nfs", rhel7_minimal]], indirect=True
-)
-def test_migration_policy_tag(request, appliance, v2v_providers, host_creds, conversion_tags,
-                              form_data_vm_obj_single_datastore, soft_assert):
-    """Test policy to prevent source VM from starting if migration is complete"""
-    infrastructure_mapping_collection = appliance.collections.v2v_mappings
-    mapping = infrastructure_mapping_collection.create(form_data_vm_obj_single_datastore.form_data)
-
-    @request.addfinalizer
-    def _cleanup():
-        infrastructure_mapping_collection.delete(mapping)
-
-    migration_plan_collection = appliance.collections.v2v_plans
-    # vm_obj is a list, with only 1 VM object, hence [0]
-    vm_obj = form_data_vm_obj_single_datastore.vm_list[0]
-    migration_plan = migration_plan_collection.create(
-        name="plan_{}".format(fauxfactory.gen_alphanumeric()),
-        description="desc_{}".format(fauxfactory.gen_alphanumeric()),
-        infra_map=mapping.name,
-        vm_list=form_data_vm_obj_single_datastore.vm_list,
-        start_migration=True,
-    )
-
-    # explicit wait for spinner of in-progress status card
-    view = appliance.browser.create_view(
-        navigator.get_class(migration_plan_collection, "All").VIEW.pick()
-    )
-    wait_for(
-        func=view.progress_card.is_plan_started,
-        func_args=[migration_plan.name],
-        message="migration plan is starting, be patient please",
-        delay=5,
-        num_sec=150,
-        handle_exception=True,
-    )
-
-    # wait until plan is in progress
-    wait_for(
-        func=view.plan_in_progress,
-        func_args=[migration_plan.name],
-        message="migration plan is in progress, be patient please",
-        delay=15,
-        num_sec=3600,
-    )
-    view.switch_to("Completed Plans")
-    view.wait_displayed()
-    migration_plan_collection.find_completed_plan(migration_plan)
-
-    logger.info("For plan {plan_name}, migration status : {count}, total time elapsed: {clock}"
-                .format(plan_name=migration_plan.name,
-                count=view.migration_plans_completed_list.get_vm_count_in_plan(migration_plan.name),
-                clock=view.migration_plans_completed_list.get_clock(migration_plan.name)))
-
-    available_tags = vm_obj.get_tags()
-    soft_assert("Migrated" in [tag.display_name for tag in available_tags])
-
-    vm_obj.wait_for_vm_state_change(desired_state=vm_obj.STATE_OFF, timeout=720)
-    vm_obj.power_control_from_cfme(option=vm_obj.POWER_ON, cancel=False)
-    view = appliance.browser.create_view(BaseLoggedInPage)
-    view.flash.assert_success_message(text="Start initiated", partial=True)
-    try:
-        vm_obj.wait_for_vm_state_change(desired_state=vm_obj.STATE_ON, timeout=120)
-    except TimedOutError:
-        pass
-    vm_state = vm_obj.find_quadicon().data["state"]
-    soft_assert(vm_state == "off")
